+++ src/js/processDOMChord.js
@@ -1,7 +1,31 @@
 const sound = require('./sound.js')
+const getNoteChord = require('./getNoteChord.js').getNoteChord
+const getNoteId = require('./getNoteId.js').getNoteId
 const updateChord = require('./updateChord.js')
 
+
+
+function preload() {
+    let notes = []
+    //new Howl();
+    //preloading notes files
+    for (i = 0; i < 12; i++) {
+        notes[i] = new Howl({
+            src: [
+                'src/sound/mp3/' + i + '.mp3',
+                'src/sound/wav/' + i + '.wav'
+            ],
+            loop: false,
+            preload: true
+        });
+        //console.log('notes:' + notes[0])
+    }
+    return notes
+}
+
 function processDOMChord(newNoteId, userChordIds) {
+
+    let notes = preload()
 
     // define bool for testing duplicate note entries
     // when key is clicked, save note in newNote
@@ -18,7 +42,7 @@ function processDOMChord(newNoteId, userChordIds) {
     // push to array if no duplicate found
     if (isDuplicate === false) {
         // play the audio
-        sound.playNote(newNoteId)
+        sound.playNote(newNoteId, notes)
 
         //push the note into the array
         userChordIds.push(newNoteId)
